refactor(gateway): split connection state from HTTP transport
