+++ src/gateway/server-connection-state.ts
@@ -1,0 +1,59 @@
+// Gateway connection and run registries.
+// This state is transport-fed but can be constructed without HTTP or WebSocket servers.
+import type { ChatAbortControllerEntry } from "./chat-abort.js";
+import { createGatewayBroadcaster } from "./server-broadcast.js";
+import {
+  createChatRunState,
+  createSessionEventSubscriberRegistry,
+  createSessionMessageSubscriberRegistry,
+} from "./server-chat-state.js";
+import type { GatewayWsClient } from "./server/ws-types.js";
+import { canReceiveSessionEvent } from "./session-sharing.js";
+
+/** Creates transport-independent connection, subscription, and run state. */
+export function createGatewayConnectionState(params: {
+  cfg: import("../config/config.js").OpenClawConfig;
+  getRuntimeConfig?: () => import("../config/config.js").OpenClawConfig;
+}) {
+  const loadRuntimeConfig = params.getRuntimeConfig ?? (() => params.cfg);
+  const clients = new Set<GatewayWsClient>();
+  const sessionEventSubscribers = createSessionEventSubscriberRegistry();
+  const sessionMessageSubscribers = createSessionMessageSubscriberRegistry();
+  const gatewayBroadcaster = createGatewayBroadcaster({
+    clients,
+    sessionMessageSubscribers,
+    canReceiveSessionEvent: (client, sessionKeys, agentId, event, payload) =>
+      canReceiveSessionEvent({
+        cfg: loadRuntimeConfig(),
+        client,
+        sessionKeys,
+        agentId,
+        event,
+        payload,
+      }),
+  });
+  const agentRunSeq = new Map<string, number>();
+  const dedupe = new Map<string, import("./server-shared.js").DedupeEntry>();
+  const chatRunState = createChatRunState();
+  const chatRunRegistry = chatRunState.registry;
+  const addChatRun = chatRunRegistry.add;
+  const removeChatRun = chatRunRegistry.remove;
+  const chatAbortControllers = new Map<string, ChatAbortControllerEntry>();
+  const chatQueuedTurns = new Map<string, import("./chat-queued-turns.js").QueuedChatTurnEntry>();
+  const toolEventRecipients = chatRunState.toolEventRecipients;
+
+  return {
+    clients,
+    ...gatewayBroadcaster,
+    agentRunSeq,
+    dedupe,
+    chatRunState,
+    addChatRun,
+    removeChatRun,
+    chatAbortControllers,
+    chatQueuedTurns,
+    toolEventRecipients,
+    sessionEventSubscribers,
+    sessionMessageSubscribers,
+  };
+}
