+++ src/gateway/test-helpers.server-runtime-state.ts
@@ -1,19 +1,20 @@
 // Server runtime-state test helper builds minimal gateway runtime state with a
 // configurable plugin registry.
 import { createEmptyPluginRegistry } from "../plugins/registry.js";
-import { createGatewayRuntimeState } from "./server-runtime-state.js";
+import { createGatewayConnectionState } from "./server-connection-state.js";
+import { createGatewayHttpTransport } from "./server-runtime-state.js";
 
 /**
  * Runtime-state fixture factory for gateway server tests.
  */
-type GatewayRuntimeStateParams = Parameters<typeof createGatewayRuntimeState>[0];
+type GatewayRuntimeStateParams = Omit<Parameters<typeof createGatewayHttpTransport>[0], "clients">;
 
 /** Creates a minimal gateway runtime state with optional plugin registry fixture. */
 export async function createGatewayRuntimeStateForTest(
   pluginRegistry: GatewayRuntimeStateParams["pluginRegistry"] = createEmptyPluginRegistry(),
   overrides: Partial<GatewayRuntimeStateParams> = {},
 ) {
-  return await createGatewayRuntimeState({
+  const params = {
     cfg: {},
     bindHost: "127.0.0.1",
     port: 0,
@@ -32,5 +33,11 @@ export async function createGatewayRuntimeStateForTest(
     logHooks: { info: () => {}, warn: () => {}, error: () => {}, debug: () => {} } as never,
     logPlugins: { info: () => {}, warn: () => {}, error: () => {}, debug: () => {} } as never,
     ...overrides,
+  };
+  const connectionState = createGatewayConnectionState(params);
+  const httpTransport = await createGatewayHttpTransport({
+    ...params,
+    clients: connectionState.clients,
   });
+  return { ...httpTransport, ...connectionState };
 }
